Add files via upload
Project Description
The new computer-controlled conveyor belt is an exciting project that moves and positions items on a conveyor belt with a high degree of accuracy (< 1 millimeter of error).  The proposed project will produce a new system capable of automating the movement of a wide variety of warehouse materials commonly used in manufacturing.  The computer-controlled conveyor belt has the potential to be a critical unit in factories producing a variety of products.  

Assumption:
A seven-day workweek is used for the whole year. No holidays.
An 8-hour workday or 56-hour workweek is used.  Overtime is not allowed.
The project should start on January 1 of the next year.
No splitting of activities is allowed.
No partial assignments are allowed (i.e. 50%).  All resources must be assigned 100%.
Resources of a particular type have identical capabilities and may be substituted for each other.  Hence, when working with resource type having multiple people, please create a single project resource listing (e.g. “Design”) and assign the resource a percentage value corresponding to the number of people (e.g. 200% for two people).  Similarly, when working with resource types having a single person, create a single project resource listing and assign it a value of 100% to indicate a single person.
Activity duration are fixed meaning adding resources to an activity does not decrease the duration of the activity.