Computer Controlled Conveyor belt
Project Description
The new computer-controlled conveyor belt is an exciting project that moves and positions items on a conveyor belt with a high degree of accuracy (< 1 millimeter of error).  The proposed project will produce a new system capable of automating the movement of a wide variety of warehouse materials commonly used in manufacturing.  The computer-controlled conveyor belt has the potential to be a critical unit in factories producing a variety of products.